+++ drones/src/main/java/br/com/fiap/drones/dto/DadosInicioMissao.java
@@ -12,6 +12,6 @@ public record DadosInicioMissao(
         @NotNull
         double longitudeInicioVoo,
         @NotNull
-        double altitude
+        int altitude
 ) {
 }
